feat(samples): telehealth changes to dark mode switcher
