+++ samples/telehealth/src/containers/AppView.scss
@@ -1,0 +1,98 @@
+main.app-view {
+  display: flex;
+  flex-direction: row;
+  height: 560px;
+  justify-content: center;
+  position: relative;
+
+  & > button {
+    bottom: 0;
+    color: #4AB2D9;
+    font-size: 12px;
+    left: 0;
+    margin: 20px;
+    padding: 8px;
+    position: absolute;
+    text-decoration: underline;
+
+    svg {
+      margin-right: 10px;
+    }
+  }
+
+  --accent-primary: #4AB2D9;
+  --accent-secondary: #9ACDE1;
+  --background-primary: #CBCBCB;
+  --background-secondary: #F4F4F4;
+  --text-primary: #FFFFFF;
+  --text-secondary: #3E3E3E;
+
+  --member__avatar__size: 46px;
+  --member__background: #FFFFFF;
+  --member__name__color: #000000;
+  --member__name__fontSize: 16px;
+  --member__padding: 12px 16px;
+  --member__title__color: #8F8F8F;
+  --member__title__fontSize: 14px;
+  --member__title__padding: 0;
+  --msg--hover__background: transparent;
+  --msg-input__background: #E9E9E9;
+  --msg-input__padding: 9px 18px 9px 0;
+  --msg-input__send--active__color: #7D99B5;
+  --msg-input__send__color: #7D99B5;
+  --msg-input__send--active__color: #7D99B5;
+  --msg-input__textarea__background: #FFFFFF;
+  --msg-input__textarea__borderRadius: 8px;
+  --msg-input__textarea__fontSize: 15px;
+  --msg-input__textarea__height: 40px;
+  --msg-input__textarea__padding: 11px;
+  --msg-list__background: #FFFFFF;
+  --msg__author__color: #7D99B5;
+  --msg__author__fontSize: 12px;
+  --msg__author__fontWeight: bold;
+  --msg__avatar__margin: 0 14px 0 0;
+  --msg__avatar__size: 32px;
+  --msg__bubble__background: #EAF0F6;
+  --msg__bubble__borderRadius: 8px;
+  --msg__bubble__color: #000000;
+  --msg__bubble__fontSize: 15px;
+  --msg__bubble__padding: 10px;
+  --msg__padding: 10px 9px;
+  --msg__time__color: #89A2BC;
+  --msg__time__fontSize: 12px;
+
+  .pn-msg--own {
+    --msg__avatar__display: none;
+    --msg__bubble__background: #D9F2FC;
+    --msg__content__alignItems: end;
+    --msg__flexDirection: row-reverse;
+  }
+
+  &.dark {
+    --accent-primary: #33687B;
+    --accent-secondary: #578291;
+    --background-primary: #2B2B2B;
+    --background-secondary: #464646;
+    --text-primary: #FFFFFF;
+    --text-secondary: #FFFFFF;
+
+    --member__background: #606060;
+    --member__name__color: #FFFFFF;
+    --member__title__color: #ADADAD;
+    --msg-input__background: #383838;
+    --msg-input__send__color: #FFFFFF;
+    --msg-input__send--active__color: #FFFFFF;
+    --msg-input__textarea__background: #606060;
+    --msg-input__textarea__color: #FFFFFF;
+    --msg-list__background: #606060;
+    --msg__author__color: #FFFFFF;
+    --msg__bubble__background: #525252;
+    --msg__bubble__color: #CECECE;
+    --msg__time__color: #CBCBCB;
+  }
+
+  &.dark .pn-msg--own {
+    --msg__bubble__background: #33687B;
+    --msg__bubble__color: #FFFFFF;
+  }
+}
