+++ samples/telehealth/src/containers/DoctorView.scss
@@ -3,7 +3,7 @@ main.doctor-view {
   box-shadow: 0px 0px 24px rgba(0, 0, 0, 0.18), 4px 4px 4px rgba(0, 0, 0, 0.1);
   display: flex;
   justify-content: end;
-  margin: 0 20px;
+  margin-left: 20px;
   overflow: hidden;
   width: 640px;
