+++ samples/telehealth/src/containers/PatientView.scss
@@ -3,7 +3,7 @@ main.patient-view {
   display: flex;
   flex-direction: column;
   justify-content: end;
-  margin: 0 20px;
+  margin-right: 20px;
   width: 330px;
 
   section {
